added cycletype. still unstable program with interaction.

+++ Project_1/inc/wavefunction.h
@@ -12,7 +12,7 @@ public:
     static double TrialWaveFunctionInteraction(const arma::mat &, const int &, const int &, const double &, const double &, const double &);
     static void QuantumForce(const arma::mat &, arma::mat &, const double &);
     static void NumericalQuantumForce(const arma::mat &, arma::mat &, const int &, const int &, const double &, const double &, const double &);
-    static void QuantumForceInteraction(const arma::mat &, arma::mat &, const double &, const int &, const int &, const double &, const int &);
+    static void QuantumForceInteraction(const arma::mat &, arma::mat &, const int &, const int &, const double &, const double &, const int &);
     static double DerivativePsi(const arma::mat &, const int &, const int &, const double &);
 };
